+++ packaging/crosswalk.spec
@@ -58,6 +58,7 @@ BuildRequires:  pkgconfig(nspr)
 BuildRequires:  pkgconfig(openssl)
 BuildRequires:  pkgconfig(scim)
 BuildRequires:  pkgconfig(vconf)
+BuildRequires:  pkgconfig(sensor)
 BuildRequires:  pkgconfig(x11)
 BuildRequires:  pkgconfig(xcomposite)
 BuildRequires:  pkgconfig(xcursor)
