+++ runtime/browser/ui/native_app_window_tizen.h
@@ -6,8 +6,7 @@
 #define XWALK_RUNTIME_BROWSER_UI_NATIVE_APP_WINDOW_TIZEN_H_
 
 #include "xwalk/runtime/browser/ui/native_app_window_views.h"
-#include "xwalk/runtime/browser/tizen/sensor_provider.h"
-
+#include "xwalk/tizen/mobile/sensor/sensor_provider.h"
 #include "xwalk/tizen/mobile/ui/tizen_system_indicator.h"
 
 namespace xwalk {
