+++ tizen/mobile/sensor/sensor_provider.cc
@@ -2,10 +2,10 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "xwalk/runtime/browser/tizen/sensor_provider.h"
+#include "xwalk/tizen/mobile/sensor/sensor_provider.h"
 
 #include "base/logging.h"
-#include "xwalk/runtime/browser/tizen/tizen_platform_sensor.h"
+#include "xwalk/tizen/mobile/sensor/tizen_platform_sensor.h"
 
 namespace xwalk {
