+++ tizen/mobile/sensor/sensor_provider.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef XWALK_RUNTIME_BROWSER_TIZEN_SENSOR_PROVIDER_H_
-#define XWALK_RUNTIME_BROWSER_TIZEN_SENSOR_PROVIDER_H_
+#ifndef XWALK_TIZEN_MOBILE_SENSOR_SENSOR_PROVIDER_H_
+#define XWALK_TIZEN_MOBILE_SENSOR_SENSOR_PROVIDER_H_
 
 #include <set>
 
@@ -54,4 +54,4 @@ class SensorProvider {
 
 }  // namespace xwalk
 
-#endif  // XWALK_RUNTIME_BROWSER_TIZEN_SENSOR_PROVIDER_H_
+#endif  // XWALK_TIZEN_MOBILE_SENSOR_SENSOR_PROVIDER_H_
